monolith/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
boto3==1.26.129
botocore==1.29.129
certifi==2023.5.7
cffi==1.15.1
circuitbreaker==1.4.0
cryptography==39.0.2
Django==4.2.1
django-s3-storage==0.14.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jmespath==1.0.1
mysqlclient==2.1.1
oci==2.101.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==23.1.1
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
